requirements.txt
asgiref==3.2.10
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
psycopg2==2.8.6
pycparser==2.20
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
zipp==3.1.0
